requirements.txt
streamlit>=1.20
pytesseract>=0.3.10
Pillow>=9.0.0
pdf2image>=1.16.0
